requirements.txt
Flask==2.0.1
firebase-admin==6.0.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0